+++ core/lib/multistore.ts
@@ -1,8 +1,10 @@
 import rootStore from '../store'
 import { loadLanguageAsync } from '@vue-storefront/i18n'
 import { initializeSyncTaskStorage } from './sync/task'
+import { Logger } from '@vue-storefront/core/lib/logger'
 import Vue from 'vue'
 import queryString from 'query-string'
+import merge from 'lodash-es/merge'
 import { RouterManager } from '@vue-storefront/core/lib/router-manager'
 import VueRouter, { RouteConfig, RawLocation } from 'vue-router'
 import config from 'config'
@@ -18,6 +20,7 @@ export interface LocalizedRoute {
 
 export interface StoreView {
   storeCode: string,
+  inherit?: string,
   disabled?: boolean,
   storeId: any,
   name?: string,
@@ -49,8 +52,20 @@ export function currentStoreView (): StoreView {
   return rootStore.state.storeView
 }
 
+export function getFinalStoreviewConfigInMultistoreMode (storeView: StoreView): StoreView {
+  if (storeView.inherit) {
+    if (!config.storeViews[storeView.storeCode]) {
+      Logger.error(`Storeview "${storeView.inherit}" doesn't exist!`)()
+    } else {
+      storeView = merge({}, [getFinalStoreviewConfigInMultistoreMode(config.storeViews[storeView.storeCode]), storeView])
+    }
+  }
+
+  return storeView
+}
+
 export function prepareStoreView (storeCode: string): StoreView {
-  let storeView = { // current, default store
+  let storeView: StoreView = { // current, default store
     tax: config.tax,
     i18n: config.i18n,
     elasticsearch: config.elasticsearch,
@@ -58,6 +73,7 @@ export function prepareStoreView (storeCode: string): StoreView {
     storeId: config.defaultStoreCode && config.defaultStoreCode !== '' ? config.storeViews[config.defaultStoreCode].storeId : 1
   }
   const storeViewHasChanged = !rootStore.state.storeView || rootStore.state.storeView.storeCode !== storeCode
+
   if (storeCode) { // current store code
     if ((storeView = config.storeViews[storeCode])) {
       storeView.storeCode = storeCode
@@ -67,10 +83,17 @@ export function prepareStoreView (storeCode: string): StoreView {
     storeView.storeCode = config.defaultStoreCode || ''
     rootStore.state.user.current_storecode = config.defaultStoreCode || ''
   }
+
+  if (storeViewHasChanged) {
+    storeView = getFinalStoreviewConfigInMultistoreMode(storeView)
+  }
+
   loadLanguageAsync(storeView.i18n.defaultLocale)
+
   if (storeViewHasChanged) {
     rootStore.state.storeView = storeView
   }
+
   if (storeViewHasChanged || Vue.prototype.$db.currentStoreCode !== storeCode) {
     if (typeof Vue.prototype.$db === 'undefined') {
       Vue.prototype.$db = {}